+++ include/igraph_centrality.h
@@ -123,10 +123,10 @@ IGRAPH_EXPORT igraph_error_t igraph_eigenvector_centrality(const igraph_t *graph
                                                            igraph_arpack_options_t *options);
 
 IGRAPH_EXPORT igraph_error_t igraph_hub_and_authority_scores(const igraph_t *graph, igraph_vector_t *hub_vector,
-                                         igraph_vector_t *authority_vector,
-                                         igraph_real_t *value, igraph_bool_t scale,
-                                         const igraph_vector_t *weights,
-                                         igraph_arpack_options_t *options);
+                                                             igraph_vector_t *authority_vector,
+                                                             igraph_real_t *value,
+                                                             const igraph_vector_t *weights,
+                                                             igraph_arpack_options_t *options);
 
 IGRAPH_EXPORT igraph_error_t igraph_constraint(const igraph_t *graph, igraph_vector_t *res,
                                     igraph_vs_t vids, const igraph_vector_t *weights);
@@ -183,17 +183,6 @@ IGRAPH_EXPORT igraph_error_t igraph_centralization_eigenvector_centrality_tmax(c
                                                                                igraph_integer_t nodes,
                                                                                igraph_neimode_t mode,
                                                                                igraph_real_t *res);
-
-/* Deprecated functions: */
-
-IGRAPH_DEPRECATED IGRAPH_EXPORT igraph_error_t igraph_hub_score(const igraph_t *graph, igraph_vector_t *vector,
-                                   igraph_real_t *value, igraph_bool_t scale,
-                                   const igraph_vector_t *weights,
-                                   igraph_arpack_options_t *options);
-IGRAPH_DEPRECATED IGRAPH_EXPORT igraph_error_t igraph_authority_score(const igraph_t *graph, igraph_vector_t *vector,
-                                         igraph_real_t *value, igraph_bool_t scale,
-                                         const igraph_vector_t *weights,
-                                         igraph_arpack_options_t *options);
 
 __END_DECLS