+++ src/centrality/hub_authority.c
@@ -23,10 +23,10 @@
 #include "igraph_adjlist.h"
 #include "igraph_interface.h"
 #include "igraph_structural.h"
-#include "igraph_blas.h"
 
 #include "centrality/centrality_internal.h"
 
+#include <float.h>
 #include <limits.h>
 
 /* struct for the unweighted variant of the HITS algorithm */
@@ -44,6 +44,45 @@ typedef struct igraph_i_kleinberg_data2_t {
     igraph_vector_t *tmp;
     const igraph_vector_t *weights;
 } igraph_i_kleinberg_data2_t;
+
+/* Checks if at least a certain fraction of HITS centrality scores are zero.
+ * Any zero value indicates that the graphs corresponding to A A^T and A^T A are
+ * not connected, and therefore the solution is not unique. However, this situation
+ * is fairly common, and difficult to control. Thefore we only warn if the number
+ * of zero values exceeds a certain fraction.
+ *
+ * To account for numerical inaccuracies, a threshold of 'eps' is used when testing for zero.
+ * This function is intended to be used with centrality values scaled such that
+ * the maximum is 1. 'eps' is chosen accordinly.
+ *
+ * See the analogous function used in igraph_eigenvector_centrality() for details
+ * on the choice of 'eps'.
+ */
+static void warn_zero_entries(const igraph_vector_t *cent) {
+    const igraph_real_t tol = 10 * DBL_EPSILON;
+    const igraph_real_t frac = 0.3; /* warn if at least this fraction of centralities is zero */
+    const igraph_integer_t n = igraph_vector_size(cent);
+
+    /* Skip check for small graphs */
+    if (n < 10) {
+        return;
+    }
+
+    const igraph_integer_t max_zero_cnt = (igraph_integer_t) frac*n;
+    igraph_integer_t zero_cnt = 0;
+
+    for (igraph_integer_t i=0; i < n; i++) {
+        igraph_real_t x = VECTOR(*cent)[i];
+        if (-tol < x && x < tol) {
+            if (++zero_cnt > max_zero_cnt) {
+                IGRAPH_WARNING(
+                    "More than a fraction %g of hub or authority scores is zero. The presence of zero values "
+                    "indicates that the solution is not unique, thus the returned result may not be meaningful.");
+                return;
+            }
+        }
+    }
+}
 
 static igraph_error_t igraph_i_kleinberg_unweighted_hub_to_auth(
         igraph_integer_t n, igraph_vector_t *to, const igraph_real_t *from,
@@ -167,6 +206,9 @@ static igraph_error_t igraph_i_kleinberg_weighted(igraph_real_t *to,
  * \ref igraph_eigenvector_centrality().
  *
  * </para><para>
+ * Results are scaled so that the largest hub and authority scores are both 1.
+ *
+ * </para><para>
  * See the following reference on the meaning of this score:
  * J. Kleinberg. Authoritative sources in a hyperlinked
  * environment. \emb Proc. 9th ACM-SIAM Symposium on Discrete
@@ -183,8 +225,6 @@ static igraph_error_t igraph_i_kleinberg_weighted(igraph_real_t *to,
  *    stored here. If a null pointer then it is ignored.
  * \param value If not a null pointer then the eigenvalue
  *    corresponding to the calculated eigenvectors is stored here.
- * \param scale If not zero then the result will be scaled such that
- *     the absolute value of the maximum centrality is one.
  * \param weights A null pointer (meaning no edge weights), or a vector
  *     giving the weights of the edges.
  * \param options Options to ARPACK. See \ref igraph_arpack_options_t
@@ -203,7 +243,7 @@ static igraph_error_t igraph_i_kleinberg_weighted(igraph_real_t *to,
  */
 igraph_error_t igraph_hub_and_authority_scores(const igraph_t *graph,
         igraph_vector_t *hub_vector, igraph_vector_t *authority_vector,
-        igraph_real_t *value, igraph_bool_t scale,
+        igraph_real_t *value,
         const igraph_vector_t *weights, igraph_arpack_options_t *options) {
 
     igraph_adjlist_t inadjlist, outadjlist;
@@ -221,7 +261,7 @@ igraph_error_t igraph_hub_and_authority_scores(const igraph_t *graph,
     if (igraph_ecount(graph) == 0) {
         /* special case: empty graph */
         if (value) {
-            *value = igraph_ecount(graph) ? 1.0 : IGRAPH_NAN;
+            *value = 0;
         }
         if (hub_vector) {
             IGRAPH_CHECK(igraph_vector_resize(hub_vector, no_of_nodes));
@@ -230,6 +270,9 @@ igraph_error_t igraph_hub_and_authority_scores(const igraph_t *graph,
         if (authority_vector) {
             IGRAPH_CHECK(igraph_vector_resize(authority_vector, no_of_nodes));
             igraph_vector_fill(authority_vector, 1.0);
+        }
+        if (no_of_nodes > 1) {
+            IGRAPH_WARNING("The graph has no edges. Hub and authortiy scores are not meaningful.");
         }
         return IGRAPH_SUCCESS;
     }
@@ -251,7 +294,7 @@ igraph_error_t igraph_hub_and_authority_scores(const igraph_t *graph,
         if (min == 0 && max == 0) {
             /* special case: all weights are zeros */
             if (value) {
-                *value = IGRAPH_NAN;
+                *value = 0;
             }
             if (hub_vector) {
                 IGRAPH_CHECK(igraph_vector_resize(hub_vector, no_of_nodes));
@@ -261,12 +304,13 @@ igraph_error_t igraph_hub_and_authority_scores(const igraph_t *graph,
                 IGRAPH_CHECK(igraph_vector_resize(authority_vector, no_of_nodes));
                 igraph_vector_fill(authority_vector, 1);
             }
+            IGRAPH_WARNING("All edge weights are zero. Hub and authortiy scores are not meaningful.");
             return IGRAPH_SUCCESS;
         }
     }
 
     if (no_of_nodes > INT_MAX) {
-        IGRAPH_ERROR("Graph has too many vertices for ARPACK", IGRAPH_EOVERFLOW);
+        IGRAPH_ERROR("Graph has too many vertices for ARPACK.", IGRAPH_EOVERFLOW);
     }
 
     if (!options) {
@@ -317,7 +361,6 @@ igraph_error_t igraph_hub_and_authority_scores(const igraph_t *graph,
                                            options, 0, &values, &vectors));
     }
 
-
     if (value) {
         *value = VECTOR(values)[0];
     }
@@ -342,7 +385,7 @@ igraph_error_t igraph_hub_and_authority_scores(const igraph_t *graph,
                 which = i;
             }
         }
-        if (scale && amax != 0) {
+        if (amax != 0) {
             igraph_vector_scale(my_hub_vector_p, 1 / VECTOR(*my_hub_vector_p)[which]);
         } else if (igraph_i_vector_mostly_negative(my_hub_vector_p)) {
             igraph_vector_scale(my_hub_vector_p, -1.0);
@@ -354,6 +397,8 @@ igraph_error_t igraph_hub_and_authority_scores(const igraph_t *graph,
                 VECTOR(*my_hub_vector_p)[i] = 0;
             }
         }
+
+        warn_zero_entries(my_hub_vector_p);
     }
 
     if (options->info) {
@@ -364,7 +409,6 @@ igraph_error_t igraph_hub_and_authority_scores(const igraph_t *graph,
     IGRAPH_FINALLY_CLEAN(2);
 
     if (authority_vector) {
-        igraph_real_t norm;
         IGRAPH_CHECK(igraph_vector_resize(authority_vector, no_of_nodes));
         igraph_vector_null(authority_vector);
         if (weights == NULL) {
@@ -372,12 +416,8 @@ igraph_error_t igraph_hub_and_authority_scores(const igraph_t *graph,
         } else {
             igraph_i_kleinberg_weighted_hub_to_auth(no_of_nodes, authority_vector, &VECTOR(*my_hub_vector_p)[0], &ininclist, graph, weights);
         }
-        if (!scale) {
-            norm = 1.0 / igraph_blas_dnrm2(authority_vector);
-        } else {
-            norm = 1.0 / igraph_vector_max(authority_vector);
-        }
-        igraph_vector_scale(authority_vector, norm);
+
+        igraph_vector_scale(authority_vector, 1.0 / igraph_vector_max(authority_vector));
     }
 
     if (!hub_vector && authority_vector) {
@@ -397,106 +437,4 @@ igraph_error_t igraph_hub_and_authority_scores(const igraph_t *graph,
     IGRAPH_FINALLY_CLEAN(1);
 
     return IGRAPH_SUCCESS;
-}
-
-/**
- * \function igraph_hub_score
- * \brief Kleinberg's hub scores.
- *
- * \deprecated-by igraph_hub_and_authority_scores 0.10.5
- *
- * The hub scores of the vertices are defined as the principal
- * eigenvector of <code>A A^T</code>, where <code>A</code> is the adjacency
- * matrix of the graph, <code>A^T</code> is its transposed.
- *
- * </para><para>
- * See the following reference on the meaning of this score:
- * J. Kleinberg. Authoritative sources in a hyperlinked
- * environment. \emb Proc. 9th ACM-SIAM Symposium on Discrete
- * Algorithms, \eme 1998. Extended version in \emb Journal of the
- * ACM \eme 46(1999). Also appears as IBM Research Report RJ 10076, May
- * 1997.
- *
- * \param graph The input graph. Can be directed and undirected.
- * \param vector Pointer to an initialized vector, the result is
- *    stored here. If a null pointer then it is ignored.
- * \param value If not a null pointer then the eigenvalue
- *    corresponding to the calculated eigenvector is stored here.
- * \param scale If not zero then the result will be scaled such that
- *     the absolute value of the maximum centrality is one.
- * \param weights A null pointer (=no edge weights), or a vector
- *     giving the weights of the edges.
- * \param options Options to ARPACK. See \ref igraph_arpack_options_t
- *    for details. Note that the function overwrites the
- *    <code>n</code> (number of vertices) parameter and
- *    it always starts the calculation from a non-random vector
- *    calculated based on the degree of the vertices.
- * \return Error code.
- *
- * Time complexity: depends on the input graph, usually it is O(|V|),
- * the number of vertices.
- *
- * \sa \ref igraph_hub_and_authority_scores() to compute
- * hub and authrotity scores efficiently at the same time,
- * \ref igraph_authority_score() for the companion measure,
- * \ref igraph_pagerank(), \ref igraph_personalized_pagerank(),
- * \ref igraph_eigenvector_centrality() for similar measures.
- */
-
-igraph_error_t igraph_hub_score(const igraph_t *graph, igraph_vector_t *vector,
-                     igraph_real_t *value, igraph_bool_t scale,
-                     const igraph_vector_t *weights,
-                     igraph_arpack_options_t *options) {
-    return igraph_hub_and_authority_scores(graph, vector, NULL, value, scale, weights, options);
-}
-
-/**
- * \function igraph_authority_score
- * \brief Kleinberg's authority scores.
- *
- * \deprecated-by igraph_hub_and_authority_scores 0.10.5
- *
- * The authority scores of the vertices are defined as the principal
- * eigenvector of <code>A^T A</code>, where <code>A</code> is the adjacency
- * matrix of the graph, <code>A^T</code> is its transposed.
- *
- * </para><para>
- * See the following reference on the meaning of this score:
- * J. Kleinberg. Authoritative sources in a hyperlinked
- * environment. \emb Proc. 9th ACM-SIAM Symposium on Discrete
- * Algorithms, \eme 1998. Extended version in \emb Journal of the
- * ACM \eme 46(1999). Also appears as IBM Research Report RJ 10076, May
- * 1997.
- *
- * \param graph The input graph. Can be directed and undirected.
- * \param vector Pointer to an initialized vector, the result is
- *    stored here. If a null pointer then it is ignored.
- * \param value If not a null pointer then the eigenvalue
- *    corresponding to the calculated eigenvector is stored here.
- * \param scale If not zero then the result will be scaled such that
- *     the absolute value of the maximum centrality is one.
- * \param weights A null pointer (=no edge weights), or a vector
- *     giving the weights of the edges.
- * \param options Options to ARPACK. See \ref igraph_arpack_options_t
- *    for details. Note that the function overwrites the
- *    <code>n</code> (number of vertices) parameter and
- *    it always starts the calculation from a non-random vector
- *    calculated based on the degree of the vertices.
- * \return Error code.
- *
- * Time complexity: depends on the input graph, usually it is O(|V|),
- * the number of vertices.
- *
- * \sa \ref igraph_hub_and_authority_scores() to compute
- * hub and authrotity scores efficiently at the same time,
- * \ref igraph_hub_score() for the companion measure,
- * \ref igraph_pagerank(), \ref igraph_personalized_pagerank(),
- * \ref igraph_eigenvector_centrality() for similar measures.
- */
-
-igraph_error_t igraph_authority_score(const igraph_t *graph, igraph_vector_t *vector,
-                           igraph_real_t *value, igraph_bool_t scale,
-                           const igraph_vector_t *weights,
-                           igraph_arpack_options_t *options) {
-    return igraph_hub_and_authority_scores(graph, NULL, vector, value, scale, weights, options);
 }